src/methods/requirements.txt
pytorch
opencv-python
tensorboard
vizdoom